♻️(referentiel) sépare IOffersRepository en base et IIngestionOffersRepository
Déplace les méthodes ingestion-spécifiques (upsert_batch, get_pending_processing,
mark_as_processed, mark_as_pending, mark_as_archived, count_published, count_archived)
vers une nouvelle interface IIngestionOffersRepository dans domain/ingestion.
IOffersRepository ne conserve que les lectures cross-domaine.

Remplace aussi le test d'intégration de CalculateDailyStatsUseCase par un test unitaire
avec mocks, conformément au pattern des tests du layer application.
