+++ libs/referentiel/repositories/offers_repository_interface.py
@@ -1,29 +1,12 @@
-from typing import List, Protocol, TypedDict
+from typing import List, Protocol
 from uuid import UUID
 
 from ddd.page_interface import IPage
 
 from referentiel.entities.offer import Offer
-from referentiel.types import IUpsertResult
-
-
-class IArchiveError(TypedDict):
-    entity_id: UUID
-    error: str
-    exception: Exception
-
-
-class IArchiveResult(TypedDict):
-    fetched: int
-    vector_deleted: int
-    entity_archived: int
-    errors: List[IArchiveError]
 
 
 class IOffersRepository(Protocol):
-    # todo move this logic in ingestion
-    def upsert_batch(self, offers_list: List[Offer]) -> IUpsertResult: ...
-
     def get_by_id(self, offer_id: UUID) -> Offer: ...
 
     def get_by_ids(self, offer_ids: List[UUID]) -> List[Offer]: ...
@@ -43,14 +26,3 @@ class IOffersRepository(Protocol):
     ) -> IPage[Offer]: ...
 
     def get_by_source_id(self, source_id: UUID) -> IPage[Offer]: ...
-
-    # todo move this logic in ingestion
-    def get_pending_processing(self, limit: int = 1000) -> List[Offer]: ...
-
-    # todo move this logic in ingestion
-    def mark_as_processed(self, offers_list: List[Offer]) -> int: ...
-
-    # todo move this logic in ingestion
-    def mark_as_pending(self, offers_list: List[Offer]) -> int: ...
-
-    def mark_as_archived(self, offers_list: List[Offer]) -> int: ...
