+++ src/web/application/ingestion/usecases/archive_offers.py
@@ -3,12 +3,12 @@ from datetime import datetime
 from asgiref.sync import async_to_sync
 from ddd.services.logger_interface import ILogger
 from ddd.usecase_interface import IUseCase
-from referentiel.repositories.offers_repository_interface import (
-    IArchiveResult,
-    IOffersRepository,
-)
 
 from domain.ingestion.gateways.document_gateway_interface import IDocumentGateway
+from domain.ingestion.repositories.ingestion_offers_repository_interface import (
+    IArchiveResult,
+    IIngestionOffersRepository,
+)
 from domain.ingestion.repositories.vector_repository_interface import IVectorRepository
 
 MAX_OFFSET = 100000
@@ -17,7 +17,7 @@ MAX_OFFSET = 100000
 class ArchiveOffersUsecase(IUseCase[datetime, IArchiveResult]):
     def __init__(
         self,
-        offers_repository: IOffersRepository,
+        offers_repository: IIngestionOffersRepository,
         document_gateway: IDocumentGateway,
         vector_repository: IVectorRepository,
         logger: ILogger,
